+++ src/vision/mobilenet.rs
@@ -73,8 +73,8 @@ pub fn v2(p: &nn::Path, nclasses: i64) -> impl ModuleT {
         .add(nn::linear(&c_p / 1, 1280, nclasses, Default::default()));
     nn::func_t(move |xs, train| {
         xs.apply_t(&features, train)
-            .mean_dim(&[2], false, crate::Kind::Float)
-            .mean_dim(&[2], false, crate::Kind::Float)
+            .mean_dim(&[2], false, crate::Kind::Double)
+            .mean_dim(&[2], false, crate::Kind::Double)
             .apply_t(&classifier, train)
     })
 }
